morning-packet/requirements.txt
google-api-python-client>=2.0.0
google-auth-httplib2>=0.1.0
google-auth-oauthlib>=1.0.0
anthropic>=0.40.0
twilio>=8.0.0
pyyaml>=6.0
python-dotenv>=1.0.0
pytz>=2023.3
requests>=2.28.0
